requirements.txt
astroid==2.0.4
Click==7.0
colorama==0.4.0
Flask==1.0.2
Flask-WTF==0.14.2
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
Pillow==5.3.0
pylint==2.1.1
six==1.11.0
typed-ast==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
